+++ docs/changelog.md
@@ -8,16 +8,19 @@ v0.6.0 (, 2023)
 * Enhancements
     * Add pyarrow dependency and use pyarrow backed dtypes [#120][#120]
     * Add Airbnb Reviews dataset [#125][#125]
+    * Add Store dataset [#131][#131]
 * Fixes
     * Rename `_execute_operations_on_df` to `target` in executed prediction problem dataframe [#124][#124]
     * Clean up operation description generation [#118][#118]
     * Remove PredictionProblemEvaluator [#118][#118]
     * Remove FeaturetoolsWrapper class [#100][#100]
+    * Remove covid19 and youtube datasets [#131][#131]
 
     [#124]: <https://github.com/trane-dev/Trane/pull/124>
     [#118]: <https://github.com/trane-dev/Trane/pull/118>
     [#120]: <https://github.com/trane-dev/Trane/pull/120>
     [#125]: <https://github.com/trane-dev/Trane/pull/125>
+    [#131]: <https://github.com/trane-dev/Trane/pull/131>
 
 
 v0.5.0 (July 27, 2023)